% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_slope_detection_regression.R
\name{CalcDist}
\alias{CalcDist}
\title{Calculates the distance of all points along the river to get the x-dimension of the regression.}
\usage{
CalcDist(dat, centerLineLL, myProj)
}
\arguments{
\item{dat}{Data coordinates in UTM}

\item{centerLineLL}{Coordinates of SWORD centerline in degrees (lon,lat)}

\item{myProj}{String containing UTM projection. Is determined based on lat and lon.}
}
\value{
Standard data frame now including the position along the river relative to arbitrary point
}
\description{
Calculates the distance of all points along the river to get the x-dimension of the regression.
}
