% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_slope_detection_regression.R
\name{Check_data_format}
\alias{Check_data_format}
\title{Checks data formats and give warnings if data is not as expected. Assigns column names to data frame}
\usage{
Check_data_format(dat)
}
\arguments{
\item{dat}{Main data frame containing satellite and SWORD data}
}
\value{
Data frame with columns names
}
\description{
Checks data formats and give warnings if data is not as expected. Assigns column names to data frame
}
