% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_slope_detection_regression.R
\name{Filter_data}
\alias{Filter_data}
\title{Outlier filtering for each SWORD node. Removes global and contextual outliers using quantiles and Gaussian kernel distribution.}
\usage{
Filter_data(dat, myProj)
}
\arguments{
\item{dat}{Main data frame containing satellite and SWORD data}

\item{myProj}{String containing UTM projection. Is determined based on lat and lon.}
}
\value{
Main data frame containing outlier filtered data.
}
\description{
Outlier filtering for each SWORD node. Removes global and contextual outliers using quantiles and Gaussian kernel distribution.
}
