% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_slope_detection_regression.R
\name{Find_slope}
\alias{Find_slope}
\title{Find_slope
Checks input data, call all other functions and saves output data to file}
\usage{
Find_slope(
  Paths,
  file2,
  SWORD,
  Max_reg_dist,
  Min_reg_dist,
  Min_reg_p,
  Occ_thr,
  filelist,
  output_file
)
}
\arguments{
\item{Paths}{list of paths containing paths from user input to ICE2WSS}

\item{file2}{Number. File number in satellite input data folder}

\item{SWORD}{list. List containing the SWORD version and SWORD area}

\item{Max_reg_dist}{A number. Maximum distance along river accepted for slope calculation}

\item{Min_reg_dist}{A number. Minimum distance along river accepted for slope calculation}

\item{Min_reg_p}{A number. Minimum number of water surface elevation points needed for slope calculation}

\item{Occ_thr}{A number. Minimum water surface occurrence accepted}

\item{filelist}{List of numbers describing files to be processed}

\item{output_file}{Name of output file. Created automatically by the script}
}
\value{
Nothing. File is produced in output path
}
\description{
Find_slope
Checks input data, call all other functions and saves output data to file
}
