% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_slope_detection_regression.R
\name{getdist}
\alias{getdist}
\title{Calculate distance between points}
\usage{
getdist(v, w)
}
\arguments{
\item{v}{Point coordinates}

\item{w}{input coordinate}
}
\value{
Distance
}
\description{
Calculate distance between points
}
