% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_slope_detection_regression.R
\name{handle_SWORD}
\alias{handle_SWORD}
\title{Loads SWORD centerlines and nodes as a text file. If this doesn't exist, the file is created based on the nc SWORD file}
\usage{
handle_SWORD(SWORD_dir, Version, Area, Log_file)
}
\arguments{
\item{SWORD_dir}{String. Path to folder with SWORD data either as text file or as nc}

\item{Version}{String. String containing SWORD version to use}

\item{Area}{String. String containing SWORD area to use}

\item{Log_file}{String. Timestamp for log file}
}
\value{
Nothing. Data is loaded
}
\description{
Loads SWORD centerlines and nodes as a text file. If this doesn't exist, the file is created based on the nc SWORD file
}
