% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_WSE_for_slope_detection.R
\name{ICE2WSS}
\alias{ICE2WSS}
\title{Process all files specified by use in parallel.}
\usage{
ICE2WSS(Paths, Files, SWORD, Max_reg_dist, Min_reg_dist, Min_reg_p, Occ_thr)
}
\arguments{
\item{Paths}{List of 3 paths: c(Outdir, ATL13_data_path, SWORD_dir)}

\item{Files}{List of files to process.}

\item{SWORD}{list. List containing the SWORD version and SWORD area}

\item{Max_reg_dist}{A number. Maximum distance along river accepted for slope calculation}

\item{Min_reg_dist}{A number. Minimum distance along river accepted for slope calculation}

\item{Min_reg_p}{A number. Minimum number of water surface elevation points needed for slope calculation}

\item{Occ_thr}{A number. Minimum water surface occurrence accepted. See https://global-surface-water.appspot.com/}
}
\value{
Nothing. File is produced in output path
}
\description{
Process all files specified by use in parallel.
}
