% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Handle_SWOT_dat.R
\name{Produce_SWORD_data}
\alias{Produce_SWORD_data}
\title{Process all files specified by use in parallel.}
\usage{
Produce_SWORD_data(SWORD_dir, Area, Version)
}
\arguments{
\item{SWORD_dir}{Path pointing to folder with SWORD data.}

\item{Area}{String. Contains area of interest of SWORD data}

\item{Version}{String. SWORD data}
}
\value{
Nothing. File is produced in SWORD subfolder
}
\description{
Process all files specified by use in parallel.
}
