% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_slope_detection_regression.R
\name{Project}
\alias{Project}
\title{Projects satellite data onto the centerline}
\usage{
Project(dat)
}
\arguments{
\item{dat}{Main data frame}
}
\value{
Reprojected data
}
\description{
Projects satellite data onto the centerline
}
