% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_slope_detection_regression.R
\name{Project_angle}
\alias{Project_angle}
\title{Calculates angle between satellite beam and local river reach}
\usage{
Project_angle(dat)
}
\arguments{
\item{dat}{Main data frame}
}
\value{
Intersection angle
}
\description{
Calculates angle between satellite beam and local river reach
}
