% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_slope_detection_regression.R
\name{project_to_centerline}
\alias{project_to_centerline}
\title{Projects satellite data onto river centrline based on nearby SWORD nodes.}
\usage{
project_to_centerline(dat)
}
\arguments{
\item{dat}{Main data frame containing satellite and SWORD data}
}
\value{
Main data frame including coordinates for the projected position and the intersection angle between river and satellite
}
\description{
Projects satellite data onto river centrline based on nearby SWORD nodes.
}
