% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_slope_detection_regression.R
\name{Find_slope}
\alias{Find_slope}
\title{Find_slope
Checks input data, call all other functions and saves output data to file}
\usage{
Find_slope(file2)
}
\arguments{
\item{file2}{Number. File number in satellite input data folder.}
}
\value{
Nothing. File is produced in output path
Find_slope()
}
\description{
Find_slope
Checks input data, call all other functions and saves output data to file
}
